pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Alexander <alexzhur87@mail.ru>"]
readme = "README.md"
keywords = ["gendiff"]
packages = [{ include = "gendiff" }]

[tool.poetry.dependencies]
python = "^3.9"
pytest-cov = "^4.1.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff ="gendiff.scripts.gendiff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
